requirements.txt
absl-py
attrs
dataclasses
imutils
mediapipe==0.8.9
numpy
opencv-python
protobuf
PyYAML
